Splits behavior for visualising Ecore metamodels and models.
This is done to improve maintainability. Common behavior is moved to a common super class EMoflonEcoreVisualiser, whereas
metamodel or model specific code is moved to the according subclass EMoflonMetamodelVisualiser and EMoflonModelVisualiser.
Additional, more precise code for PlantUML API interaction has been added, to determine as precise as possible, whether a given
selection is handled by the metamodel or model visualiser, or not at all. Ecore visualisers now should act as intended by the
PlantUML API.

Also removes code to handle correspondence model. Firstly, this functionality does not belong with functionality to visualise
Ecore metamodels or models. Secondly, the improved checking methods for supported editors and selections allow for a more
appropriate visualiser to handle correspondence models.
