+++ org.moflon.core.ui/plugin.xml
@@ -240,7 +240,9 @@
     </factory>
   </extension>
   <extension point="net.sourceforge.plantuml.eclipse.diagramTextProvider">
-    <diagramTextProvider priority="normal" providerClass="org.moflon.core.ui.visualisation.EMoflonModelAndMetamodelVisualiser">
-     </diagramTextProvider>
+    <diagramTextProvider priority="normal" providerClass="org.moflon.core.ui.visualisation.EMoflonModelVisualiser">
+    </diagramTextProvider>
+    <diagramTextProvider priority="normal" providerClass="org.moflon.core.ui.visualisation.EMoflonMetamodelVisualiser">
+    </diagramTextProvider>
   </extension>
 </plugin>
